# [Perf] Convert np array to torch tensor to index into block table for attn chunking